+++ lib/workos.rb
@@ -27,6 +27,7 @@ module WorkOS
 
   autoload :Types, 'workos/types'
   autoload :Base, 'workos/base'
+  autoload :Client, 'workos/client'
   autoload :AuditLog, 'workos/audit_log'
   autoload :Profile, 'workos/profile'
   autoload :SSO, 'workos/sso'
